+++ src/redux/features/form/formSlice.tsx
@@ -1,5 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { FormValues } from 'types/types';
+import { FormValues } from 'types';
 import { RootState } from '../../store/store';
 
 const initialState: FormValues = {
