+++ src/redux/store/index.tsx
@@ -1,0 +1,3 @@
+import { store } from './store';
+
+export default store;
